refactor(runtime): split transcript page mappers
Summary:
- split ticket run transcript page mapping into focused frontend modules
- keep transcript data exports wired through the new page and entry mappers
- clean runtime projection and API test lint issues found by the CI gate

Rationale:
- keep the ASE-160 transcript reader changes maintainable while preserving the
  new raw/activity/transcript UI loading path
- unblock repository CI by removing dead projection helpers and staticcheck
  issues in the new runtime event model work

Tests:
- PATH=$HOME/.nvm/versions/node/v22.22.1/bin:$HOME/.local/go1.26.1/bin:$PATH OPENASE_PGTEST_SHARED_ROOT=$HOME/.cache/openase/pgtest .codex/skills/push/scripts/openase_ci_gate.sh
